+++ src/functions/mainArgs.js
@@ -1,6 +1,0 @@
-const _mainArgs = () => {
-  const args = process.argv;
-  return args.splice(2);
-};
-
-module.exports = _mainArgs;