+++ src/functions/multiParam.js
@@ -1,8 +1,0 @@
-const _multiParam = (command) => {
-  const params = command.split(" ");
-
-  params.shift();
-  return params;
-};
-
-module.exports = _multiParam;